+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 
+## [0.3.0] - 2018-04-06
+### Added
+- Added clippy checks without using it as an optional dependency
+
+### Modified
+- Changed the error handling from [error-chain](https://crates.io/crates/error-chain)
+to [failure](https://crates.io/crates/failure)
+
+### Removed
+- Removed clippy as an optional dependency
+
 ## [0.2.0] - 2017-02-04
 ### Added
 - Now works on stable
@@ -18,6 +29,7 @@ crate to generate the fixtures
 
 ## Removed
 - Removed fixtures
+
 ## [0.1.1] - 2016-12-25
 ### Added
 - Clippy lints